brought in old changes from previous turnbased branch
