+++ code/beastly-brawl-showdown/imports/data/monsters/MonsterData.ts
@@ -1,0 +1,9 @@
+import MysticWyvern from './MysticWyvern';
+import ShadowFangPredator from './ShadowFangPredator';
+import StoneHideGuardian from './StoneHideGuardian';
+
+export const mysticWyvern = new MysticWyvern();
+export const shadowFang = new ShadowFangPredator();
+export const stoneHide = new StoneHideGuardian();
+
+export const allMonsters = [mysticWyvern, shadowFang, stoneHide];
