+++ code/beastly-brawl-showdown/imports/data/monsters/MysticWyvern.ts
@@ -1,5 +1,23 @@
-import Monsters from './Monsters';
-import DiceRoller from '../utils/DiceRoller';
+import Monsters from "./Monsters";
+import DiceRoller from "../utils/DiceRoller";
+import readline from "readline";
+
+/**
+ * Prompts the user to answer yes or no.
+ */
+function askYesNo(question: string): Promise<boolean> {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  return new Promise((resolve) => {
+    rl.question(`${question} (y/n): `, (answer) => {
+      rl.close();
+      resolve(answer.trim().toLowerCase() === "y");
+    });
+  });
+}
 
 /**
  * MysticWyvern monster subclass.
@@ -10,7 +28,15 @@ export default class MysticWyvern extends Monsters {
   private rerollCharge: boolean;
 
   constructor() {
-    super(25, 14, 2, "Reroll once per battle", "Balanced");
+    super(
+      25,
+      14,
+      2,
+      "Reroll once per battle",
+      "Balanced",
+      "/img/monster-selection-images/placeholder_monster_1.png",
+      "/img/monster-image/dragon.png"
+    );
     this.rerollCharge = true;
   }
 
@@ -21,7 +47,7 @@ export default class MysticWyvern extends Monsters {
    * defender: Monsters
    * TODO:FIX IF NEEDED
    */
-  attack(): number {
+  override async attack(): Promise<number> {
     let roll = DiceRoller.d20();
     console.log(`${this.type} rolls ${roll}.`);
 
@@ -29,7 +55,9 @@ export default class MysticWyvern extends Monsters {
     if (this.rerollCharge) {
       // Note: In Node.js or non-browser environments, confirm won't work.
       // You may want to replace with a different prompt method or remove this in non-browser context.
-      const reroll = confirm("Do you want to reroll the d20?");
+      //const reroll = confirm("Do you want to reroll the d20?");
+      const reroll = await askYesNo("Do you want to reroll the d20?"); //change to read in terminal
+
       if (reroll) {
         this.rerollCharge = false;
         roll = DiceRoller.d20();
